Move files; update path
