+++ frontend/src/assets/styles/Footer.module.css
