+++ frontend/src/components/layout/Footer.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import styles from './Footer.module.css';
+import styles from '../../assets/styles/Footer.module.css';
 
 export const Footer = () => {
 	return (
