+++ frontend/src/components/layout/Main.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import styles from './Main.module.css';
+import styles from '../../assets/styles/Main.module.css';
 
 export const Main = () => {
 	return (
