+++ frontend/src/pages/Layout.jsx
@@ -1,15 +1,15 @@
-import { Footer } from '../components/Footer';
-import { Header } from '../components/Header';
-import { Main } from '../components/Main';
+import { Footer } from '../components/layout/Footer';
+import { Header } from '../components/layout/Header';
+import { Main } from '../components/layout/Main';
 
 export const Layout = () => {
-  return (
-    <>
-      <Header />
+	return (
+		<>
+			<Header />
 
-      <Main />
+			<Main />
 
-      <Footer />
-    </>
-  );
+			<Footer />
+		</>
+	);
 };
